requirements.txt
antlr4-python3-runtime==4.10
cycler==0.11.0
fonttools==4.33.3
glfw==2.5.3
kiwisolver==1.4.3
matplotlib==3.5.2
mpmath==1.2.1
numpy==1.23.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
PyOpenGL==3.1.6
pyparsing==3.0.9
PyQt6==6.3.1
PyQt6-Qt6==6.3.2
PyQt6-sip==13.4.0
python-dateutil==2.8.2
pytz==2022.1
shape==0.1.1
shapes==0.1.3
six==1.16.0
sympy==1.11.1
torch==1.12.1
typing_extensions==4.3.0
